fix build

+++ client/src/components/DeleteBook.js
@@ -5,13 +5,39 @@ class DeleteBook extends Component {
     constructor(props) {
         super();
         this.state = {
-            books: []
+            books: [],
+            id: ''
         }
     }
 
-    deleteBook = (e, id) => {
-        e.preventDefault();
-        fetch(`/books/${id}`, {
+    componentDidMount = () => {
+        fetch(`/books`, {
+            method: "GET",
+            headers: {
+                "Content-Type": "application/json"
+            }
+        })
+            .then(response => {
+                if (response.ok) {
+                    return response.json();
+                }
+                throw new Error(`Network response error: ${response.id}, ${response.message}`);
+            })
+            .then(response => this.setState({ books: response.data }))
+            .catch(error => console.log(error.message));
+    }
+
+    onChange = (event) => {
+        this.setState({ id: event.target.value })
+    }
+
+    deleteBook = (event) => {
+        event.preventDefault();
+
+        if (this.state.id.length === 0)
+            return;
+
+        fetch(`/books/${this.state.id}`, {
             method: "DELETE",
             headers: {
                 "Content-Type": "application/json"
@@ -19,12 +45,11 @@ class DeleteBook extends Component {
         })
             .then(response => {
                 if (response.ok) {
-                    const { message, data } = response.json();
-                    this.setState({ books: data });
-                    return message;
+                    return response.json();
                 }
                 throw new Error(`Network response error: ${response.id}, ${response.message}`);
             })
+            .then(response => { console.log(response); this.setState({ book: response.data }) })
             .catch(error => console.log(error.message));
     }
 
@@ -36,11 +61,15 @@ class DeleteBook extends Component {
                     <Form className="form-inline">
                         <label className="sr-only">Id</label>
                         <input type="text" className="form-control mb-2 mr-sm-2" placeholder="Id" required onChange={this.onChange} />
-                        <Button type="submit" className="btn btn-danger" onSubmit={this.deleteBook}>Delete by Id</Button>
+                        <Button type="submit" className="btn btn-danger" onClick={this.deleteBook}>Delete by Id</Button>
                     </Form>
-                    <span>
-                        {this.state.book ? <li>deleted book: {this.state.book.title}</li> : "No book deleted"}
-                    </span>
+                    <ul>
+                        {
+                            this.state.books.map((book) => {
+                                return <li key={book._id}>Book Id: {book._id}, Title: {book.title}</li>
+                            })
+                        }
+                    </ul>
                 </Card.Body>
             </Card>
         );
